add delete message logic

+++ server/lib/src/db_server/database_helper/messages/message_impl.dart
@@ -87,8 +87,7 @@ class MessagesServices implements IMessagesServices {
       {required String newValues, required String condition}) async {
     Database db = await DbServerServices.instanse.database;
     return await db
-            .rawUpdate('''UPDATE messages SET $newValues WHERE ($condition)''')
-        as int;
+        .rawUpdate('''UPDATE messages SET $newValues WHERE ($condition)''');
   }
 
   // @override
